% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapfill.R
\name{Image}
\alias{Image}
\title{Image Panels}
\usage{
Image(
  x = NULL,
  zlim = range(x, na.rm = TRUE),
  col = fields::tim.colors(1000),
  theme = TRUE,
  guides = TRUE,
  na.value = "black",
  panelsByrow = TRUE,
  asRaster = TRUE,
  xlab = "",
  ylab = "",
  colbarTitle = "",
  ...
)
}
\arguments{
\item{x}{Numeric array with 4, 3, or 2 dimensions
containing the data to be plotted.}

\item{zlim}{Numeric vector of length 2.
Gives the upper and lower bound of the plotted values.}

\item{col}{Vector of colors.}

\item{theme}{Logical vector of length one.
Should the ggplot2 theme be modified?}

\item{guides}{Logical vector of length one.
Should ggplot2 guides be modified?}

\item{na.value}{Vector of length one.
The color to be used for NA values.}

\item{panelsByrow}{Logical vector of length one.
Indicates the ordering of the panels.}

\item{asRaster}{Logical vector of length one.
If \code{TRUE}, the ordering of the pixel within images is similar to the plot method of the raster package.
Otherwise, the ordering is similar to \code{\link{image}}.}

\item{xlab}{Character vector (or expression) of length one giving the x-axis label.}

\item{ylab}{Character vector (or expression) of length one giving the y-axis label.}

\item{colbarTitle}{Character vector (or expression) of length one giving the colorbar label.}

\item{...}{Additional arguments are passed to \code{ggplot}.}
}
\value{
Object (plot) of class \code{c("gg", "ggplot2")}.
}
\description{
Creates an image panel to visualize data in 4, 3 or 2 dimensional arrays (e.g., space-time data). 
The function returns a \code{ggplot2} object, which 
can be modified using ggplot2 (and/or grid) syntax.
}
\examples{
library("abind")
t1 <- array(rep(c(1,0), each = 5), c(5,5))
t1[5,3] <- 2 
t2 <- abind(t1, t1, along = 3)
t3 <- abind(t2, t2, along = 4)
Image(t1)
Image(t2)
Image(t3)

\dontrun{
Image(ndvi)

p1 <- Image(ndvi, colbarTitle = "NDVI", xlab = "Year", ylab = "DOY",
            panelsByrow  = FALSE)
p1

p2 <- Image(ndvi[,,3,2], na.value = "white", colbarTitle = "NDVI") +
      theme(strip.text.x = element_blank(),
            strip.text.y = element_blank(),
            panel.border = element_rect(fill = NA, size = 1))
p2

## place modified color bar left
p2 + guides(fill = guide_colorbar(title = "NDVI", 
                                  barwidth = 1,
                                  barheight = 20,
                                  label.position = "right", 
                                  legend.position = c(0, 0))) +
     theme(legend.position = "right")

## place color bar at bottom
p2 + guides(fill = guide_colorbar(title = "NDVI", 
                                  barwidth = 7,
                                  barheight = .7,
                                  label.position = "bottom", 
                                  legend.position = c(0, 0)),
                                  direction = "horizontal") +
     theme(legend.position = "bottom")
}
}
\seealso{
\code{\link{ndvi}}, \code{\link[ggplot2]{ggplot2}}.
}
\author{
Florian Gerber, \email{flora.fauna.gerber@gmail.com}.
}
